# backend/src/test/resources/openapi/api-docs.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Mail Backend API",
    "description" : "REST API for the Tauri desktop mail client — accounts, messages, drafts, folders.",
    "license" : {
      "name" : "Proprietary"
    },
    "version" : "1.0.0"
  },
  "security" : [ {
    "ApiKeyAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/v1/accounts/providers" : {
      "get" : {
        "tags" : [ "Mail Providers" ],
        "summary" : "List all providers",
        "description" : "Returns the list of all predefined mail provider templates (IMAP/SMTP settings).",
        "operationId" : "getAllProviders",
        "responses" : {
          "503" : {
            "description" : "Service Unavailable"
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/MailProviderResponse"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/providers/{id}" : {
      "get" : {
        "tags" : [ "Mail Providers" ],
        "summary" : "Provider detail",
        "description" : "Returns one specific provider template by ID, or 404 if it does not exist.",
        "operationId" : "getProviderById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "exclusiveMinimum" : 0
          }
        } ],
        "responses" : {
          "503" : {
            "description" : "Service Unavailable"
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MailProviderResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/providers/resolve" : {
      "get" : {
        "tags" : [ "Mail Providers" ],
        "summary" : "Auto-detect provider by e-mail",
        "description" : "Finds the matching provider template for the given e-mail address (by domain). Returns 404 if no template matches.",
        "operationId" : "resolveProvider",
        "parameters" : [ {
          "name" : "email",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "email",
            "maxLength" : 255,
            "minLength" : 0
          }
        } ],
        "responses" : {
          "503" : {
            "description" : "Service Unavailable"
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MailProviderResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ProblemDetail" : {
        "description" : "RFC 9457 problem+json — format of all error responses.",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri",
            "description" : "URI identifier of the error type (e.g. .../errors/account_not_found)."
          },
          "title" : {
            "type" : "string",
            "description" : "Short human-readable description of the error class."
          },
          "status" : {
            "description" : "HTTP status code."
          },
          "detail" : {
            "type" : "string",
            "description" : "Error detail intended for the user."
          },
          "instance" : {
            "type" : "string",
            "format" : "uri",
            "description" : "URI of the specific occurrence (typically the request URI)."
          },
          "errorCode" : {
            "type" : "string",
            "description" : "Machine-readable error code (e.g. ACCOUNT_NOT_FOUND, VALIDATION_ERROR)."
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "Time when the error occurred (ISO-8601)."
          }
        }
      },
...
      "MailProviderResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "imapHost" : {
            "type" : "string"
          },
          "imapPort" : {
            "type" : "integer",
            "format" : "int32"
          },
          "imapSsl" : {
            "type" : "boolean"
          },
          "smtpHost" : {
            "type" : "string"
          },
          "smtpPort" : {
            "type" : "integer",
            "format" : "int32"
          },
          "smtpSsl" : {
            "type" : "boolean"
          },
          "domains" : {
            "type" : "string"
          },
          "supportsOauth2" : {
            "type" : "boolean"
          },
          "oauth2RegistrationId" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "ApiKeyAuth" : {
        "type" : "apiKey",
        "description" : "Internal API key that the desktop client loads from session.json and sends in the X-API-KEY header.",
        "name" : "X-API-KEY",
        "in" : "header"
      }
    }
  }
}